requirements.txt
django
djangorestframework
markdown
django-filter
pillow
psycopg2 